refactor: Combie type & api
