--- frontend/feature/demand-planning/service/plan-item/plan-item.api.ts
@@ -1,6 +1,0 @@
-import { request } from '../api';
-
-import { PlanItem } from './plan-item.type';
-
-export const updateOne = (data: PlanItem) =>
-    request<any>({ url: `/plan/${data.id}`, data });
