+++ frontend/feature/netsuite/class.ts
@@ -1,0 +1,9 @@
+import { request } from './api';
+
+export type Class = {
+    id: number;
+    sku: string;
+    name: string;
+};
+
+export const get = () => request<Class[]>({ url: '/class' });
