--- frontend/feature/netsuite/class/class.api.ts
@@ -1,5 +1,0 @@
-import { request } from '../api';
-
-import { Class } from './class.type';
-
-export const get = () => request<Class[]>({ url: '/class' });
