--- frontend/feature/netsuite/class/class.type.ts
@@ -1,5 +1,0 @@
-export type Class = {
-    id: number;
-    sku: string;
-    name: string;
-};
